Add Date and DateTime CFs to bulk update
Date conversion is needed only for delete because AddValueForObject
in RT::CustomField calls CanonicalizeValue methods to handle
conversion on add.

Don't pass a Default value for Date and DateTime because the
display in parens differs from the dates in the top section
which puts values in the text box like other fields. This
will ulimately be fixed for 4.4 via 4.2/date-input-keep-value.